requirements.txt
django-crispy-forms==2.1
django-celery-beat